For #668: Restyles custom tabs
