+++ app/src/main/java/org/mozilla/fenix/components/toolbar/ToolbarIntegration.kt
@@ -12,8 +12,12 @@ import mozilla.components.browser.session.runWithSession
 import mozilla.components.browser.toolbar.BrowserToolbar
 import mozilla.components.concept.storage.HistoryStorage
 import mozilla.components.feature.toolbar.ToolbarAutocompleteFeature
+import mozilla.components.feature.toolbar.ToolbarFeature
 import mozilla.components.feature.toolbar.ToolbarPresenter
+import mozilla.components.lib.publicsuffixlist.PublicSuffixList
 import mozilla.components.support.base.feature.LifecycleAwareFeature
+import org.mozilla.fenix.DefaultThemeManager
+import org.mozilla.fenix.R
 import org.mozilla.fenix.browser.BrowserFragmentDirections
 import org.mozilla.fenix.ext.components
 
@@ -27,6 +31,9 @@ class ToolbarIntegration(
     sessionId: String? = null,
     isPrivate: Boolean
 ) : LifecycleAwareFeature {
+
+    private var renderStyle: ToolbarFeature.RenderStyle = ToolbarFeature.RenderStyle.UncoloredUrl
+
     init {
         toolbar.setMenuBuilder(toolbarMenu.menuBuilder)
         toolbar.private = isPrivate
@@ -35,7 +42,10 @@ class ToolbarIntegration(
             sessionManager.runWithSession(sessionId) {
                 it.isCustomTabSession()
             }.also { isCustomTab ->
-                if (isCustomTab) return@run
+                if (isCustomTab) {
+                    renderStyle = ToolbarFeature.RenderStyle.RegistrableDomain
+                    return@run
+                }
 
                 val tabsAction = TabCounterToolbarButton(
                     sessionManager,
@@ -58,7 +68,9 @@ class ToolbarIntegration(
     private val toolbarPresenter: ToolbarPresenter = ToolbarPresenter(
         toolbar,
         context.components.core.sessionManager,
-        sessionId
+        sessionId,
+        ToolbarFeature.UrlRenderConfiguration(PublicSuffixList(context),
+            DefaultThemeManager.resolveAttribute(R.attr.primaryText, context), renderStyle = renderStyle)
     )
 
     override fun start() {
