+++ app/src/main/java/org/mozilla/fenix/components/toolbar/ToolbarUIView.kt
@@ -5,9 +5,11 @@
 package org.mozilla.fenix.components.toolbar
 
 import android.graphics.drawable.BitmapDrawable
+import android.view.Gravity
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import android.widget.ImageView
+import androidx.core.content.ContextCompat
 import io.reactivex.Observable
 import io.reactivex.Observer
 import io.reactivex.functions.Consumer
@@ -16,6 +18,7 @@ import mozilla.components.browser.toolbar.BrowserToolbar
 import mozilla.components.support.ktx.android.content.res.pxToDp
 import org.jetbrains.anko.backgroundDrawable
 import org.mozilla.fenix.R
+import org.mozilla.fenix.customtabs.CustomTabToolbarMenu
 import org.mozilla.fenix.ext.components
 import org.mozilla.fenix.mvi.UIView
 
@@ -41,6 +44,9 @@ class ToolbarUIView(
         .inflate(R.layout.layout_url_background, container, false)
 
     init {
+        val session = sessionId?.let { view.context.components.core.sessionManager.findSessionById(sessionId) }
+            ?: view.context.components.core.sessionManager.selectedSession
+
         view.apply {
             setOnUrlCommitListener {
                 actionEmitter.onNext(SearchAction.UrlCommitted(it, sessionId, state?.engine))
@@ -52,7 +58,13 @@ class ToolbarUIView(
             }
 
             browserActionMargin = resources.pxToDp(browserActionMarginDp)
-            urlBoxView = urlBackground
+
+            val isCustomTabSession = (session?.isCustomTabSession() == true)
+
+            urlBoxView = if (isCustomTabSession) { null } else urlBackground
+            progressBarGravity = if (isCustomTabSession) { Gravity.TOP } else Gravity.BOTTOM
+
+            textColor = ContextCompat.getColor(context, R.color.photonGrey30)
 
             hint = context.getString(R.string.search_hint)
 
@@ -74,17 +86,25 @@ class ToolbarUIView(
         }
 
         with(view.context) {
-            val session = sessionId?.let { components.core.sessionManager.findSessionById(sessionId) }
-                ?: components.core.sessionManager.selectedSession
+            val isCustom = session?.isCustomTabSession() ?: false
+
+            val menuToolbar = if (isCustom) {
+                CustomTabToolbarMenu(this,
+                    requestDesktopStateProvider = { session?.desktopMode ?: false },
+                    onItemTapped = { actionEmitter.onNext(SearchAction.ToolbarMenuItemTapped(it)) }
+                )
+            } else {
+                DefaultToolbarMenu(this,
+                    sessionId = sessionId,
+                    requestDesktopStateProvider = { session?.desktopMode ?: false },
+                    onItemTapped = { actionEmitter.onNext(SearchAction.ToolbarMenuItemTapped(it)) }
+                )
+            }
 
             toolbarIntegration = ToolbarIntegration(
                 this,
                 view,
-                ToolbarMenu(this,
-                    sessionId = sessionId,
-                    requestDesktopStateProvider = { session?.desktopMode ?: false },
-                    onItemTapped = { actionEmitter.onNext(SearchAction.ToolbarMenuItemTapped(it)) }
-                ),
+                menuToolbar,
                 ShippedDomainsProvider().also { it.initialize(this) },
                 components.core.historyStorage,
                 components.core.sessionManager,
